Migrated from Sourceforge
This project is quite mature, and has been on Sourceforge all this time.
This cut is a major facelift to migrate to Java 8, latest db4o and
JasperReports. Also upgrade build environment to Gradle to make it
easier for cross platform builds of SWT libraries, which are native to
each platform.